Star us on GitHub!
Star

UnrollingAverages

UnrollingAverages

Installation

Usage

How unroll_recursive works

Internals

UnrollingAverages is a Julia package aimed at reversing (or unrolling) moving averages of time series to get the original ones back.

UnrollingAverages currently assumes that the moving average is a simple moving average. Further relaxations and extensions may come in the future, see Future Developments section.

Installation

Press ] in the Julia REPL and then

pkg> add UnrollingAverages

Usage

The package exports a single function called unroll: it returns a Vector whose elements are the possible original time series.

UnrollingAverages.unroll — Method

unroll(moving_average::Vector{Float64}, window::Int64; initial_conditions::U = nothing, assert_natural::Bool = false) where { U <: Union{ Tuple{Vararg{Union{Int64,Float64}}},Nothing} }

Retrieve original time series (i.e. unroll) from its moving average moving_average.

Arguments

moving_average::Vector{Float64}: the time series representing the moving average to unroll;

window:::Int64: the width of the moving average;

initial_conditions::U = nothing: the initial values of the original time series to be recovered. It may be a Tuple of window-1 float or integer values, or nothing if initial conditions are unknown;

assert_natural::Bool = false default boolean argument. If true, then the pipeline will try to recover a time series of natural numbers only. More then one acceptable time series (where "acceptable" means that it reproduces moving_average) may be found and returned.

NB: If isnothing(initial_conditions) && !assert_natural , then only an approximate method may be used, see this StackExchange post.

A few remarks:

If isnothing(initial_conditions):

if assert_natural, then an internal unroll_iterative method is called, which tries to exactly recover the whole time series, initial conditions included. Enter ?UnrollingAverages.unroll_iterative in a julia to read details ;

if !assert_natural, then an internal unroll_linear_approximation method is called. See this StackExchange post. NB: this is an approximated method, it will generally not return the exact original time series ;

If typeof(initial_conditions) <: Ntuple{window-1, <:Union{Int64,Float64}}, then an internal unroll_recursive method is called, which exactly recovers the time series. Mathematical details about this function are reported in section How unroll_recursive works, and you may read more by entering ?UnrollingAverages.unroll_recursive.

How unroll_recursive works

This method is capable of exactly recovering the original time series $x_{t}$ (with $t \in \mathbb{N}$), given its moving average $a_t$ (with window width $W = n_{-} + n_{+} + 1$ and $t = 1,...,N−(W−1)$) and initial conditions initial_conditions (the latter must be a NTuple{W-1,Union{Int64,Float64}}):

\[a_t = \frac{1}{W}\sum_{j=t-n_-}^{t+n_+}x_j\]

By exploiting the derived recursive relation:

\[x_{t+n_+ + n_-} = W a_t - \sum\limits_{j=1}^{t+n_-+n_+-1}x_j\]